requirements.txt
chromadb==0.6.3
mcp==1.6.0
google-generativeai==0.8.4
